# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PANOR Clinical Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Patient list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientList"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
  /patients/{id}/timeline:
    get:
      operationId: getPatientTimeline
      tags: [patients]
      summary: Get patient medical timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineList"
  /patients/{id}/medications:
    get:
      operationId: getPatientMedications
      tags: [patients]
      summary: Get all active medications for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationList"

  # DOCTORS
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Doctor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorList"
  /doctors/{id}/schedule:
    get:
      operationId: getDoctorSchedule
      tags: [doctors]
      summary: Get doctor schedule and availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorSchedule"
  /doctors/{id}/schedule/slots:
    patch:
      operationId: updateDoctorSlots
      tags: [doctors]
      summary: Update doctor availability slots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSlotsBody"
      responses:
        "200":
          description: Updated slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorSchedule"

  # VISITS
...
  /lab/orders:
    get:
      operationId: listLabOrders
      tags: [lab]
      summary: List lab orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lab orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrderList"
    post:
      operationId: createLabOrder
      tags: [lab]
      summary: Create lab order from clinical notes (AI translates)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabOrderBody"
      responses:
        "201":
          description: Lab order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"
  /lab/orders/{id}:
    patch:
      operationId: updateLabOrder
      tags: [lab]
      summary: Update lab order status or results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabOrder"

  # APPOINTMENTS
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard statistics summary
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent clinical activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityList"
  /dashboard/alerts:
    get:
      operationId: getActiveAlerts
      tags: [dashboard]
      summary: Get active clinical alerts and warnings
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertList"
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        medicalId:
          type: string
        name:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        phone:
          type: string
        bloodGroup:
          type: string
          nullable: true
        allergies:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, medicalId, name, dateOfBirth, gender, phone, allergies, createdAt]
    PatientDetail:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            activeConditions:
              type: array
              items:
                type: string
            riskScore:
              type: integer
            riskLabel:
              type: string
            lastVisit:
              type: string
              nullable: true
            totalVisits:
              type: integer
    PatientList:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
      required: [patients, total]
    CreatePatientBody:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        phone:
          type: string
        bloodGroup:
          type: string
          nullable: true
        allergies:
          type: array
          items:
            type: string
      required: [name, dateOfBirth, gender, phone]
    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [visit, prescription, lab_result, alert, followup]
        title:
          type: string
        description:
          type: string
        date:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
          nullable: true
      required: [id, type, title, description, date]
    TimelineList:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
      required: [events]
    Medication:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        prescribedBy:
          type: string
        prescribedAt:
          type: string
        active:
          type: boolean
      required: [id, name, dosage, frequency, prescribedBy, prescribedAt, active]
    MedicationList:
      type: object
      properties:
        medications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
      required: [medications]
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        qualifications:
          type: string
        rating:
          type: number
        availableToday:
          type: boolean
        nextAvailable:
          type: string
          nullable: true
      required: [id, name, specialization, qualifications, rating, availableToday]
    DoctorList:
      type: object
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
      required: [doctors]
    TimeSlot:
      type: object
      properties:
        time:
          type: string
        available:
          type: boolean
        appointmentId:
          type: integer
          nullable: true
      required: [time, available]
    DoctorSchedule:
      type: object
      properties:
        doctorId:
          type: integer
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required: [doctorId, date, slots]
    UpdateSlotsBody:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required: [date, slots]
...
    LabOrder:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        visitId:
          type: integer
          nullable: true
        clinicalIntent:
          type: string
        tests:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              code:
                type: string
              priority:
                type: string
                enum: [routine, urgent, stat]
              instructions:
                type: string
            required: [name, code, priority]
        priority:
          type: string
          enum: [routine, urgent, stat]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        results:
          type: string
          nullable: true
        aiTranslation:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, patientId, patientName, clinicalIntent, tests, priority, status, createdAt]
    LabOrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/LabOrder"
        total:
          type: integer
      required: [orders, total]
    CreateLabOrderBody:
      type: object
      properties:
        patientId:
          type: integer
        visitId:
          type: integer
          nullable: true
        clinicalIntent:
          type: string
        priority:
          type: string
          enum: [routine, urgent, stat]
          default: routine
      required: [patientId, clinicalIntent]
    UpdateLabOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        results:
          type: string
...
    DashboardSummary:
      type: object
      properties:
        totalPatients:
          type: integer
        activeVisits:
          type: integer
        pendingLabOrders:
          type: integer
        todayAppointments:
          type: integer
        pendingBills:
          type: integer
        criticalAlerts:
          type: integer
        revenueThisMonth:
          type: number
        avgConsultationTime:
          type: number
        topConditions:
          type: array
          items:
            type: object
            properties:
              condition:
                type: string
              count:
                type: integer
            required: [condition, count]
      required: [totalPatients, activeVisits, pendingLabOrders, todayAppointments, pendingBills, criticalAlerts, revenueThisMonth, avgConsultationTime, topConditions]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
        severity:
          type: string
          nullable: true
      required: [id, type, description, timestamp]
    ActivityList:
      type: object
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [activities]
    Alert:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
          nullable: true
        patientName:
          type: string
          nullable: true
        type:
          type: string
          enum: [drug_interaction, cardiac_risk, diabetic_risk, outbreak, followup_overdue, critical_lab]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        isResolved:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, message, severity, isResolved, createdAt]
    AlertList:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        criticalCount:
          type: integer
      required: [alerts, criticalCount]