# lib/api-zod/src/generated/types/labOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PANOR Clinical Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { LabOrderPriority } from "./labOrderPriority";
import type { LabOrderStatus } from "./labOrderStatus";
import type { LabOrderTestsItem } from "./labOrderTestsItem";

export interface LabOrder {
  id: number;
  patientId: number;
  patientName: string;
  visitId?: number | null;
  clinicalIntent: string;
  tests: LabOrderTestsItem[];
  priority: LabOrderPriority;
  status: LabOrderStatus;
  results?: string | null;
  aiTranslation?: string | null;
  createdAt: string;
}