# lib/api-zod/src/generated/types/labOrderList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PANOR Clinical Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { LabOrder } from "./labOrder";

export interface LabOrderList {
  orders: LabOrder[];
  total: number;
}